requirements.txt
# Requirements Python pour le Système Qualiopi
# Date: 8 octobre 2025
# Version: 1.0

# Base de données
supabase==2.0.3

# Génération de documents
python-docx==1.1.0
reportlab==4.0.7

# Emails
resend==0.6.0

# Utilitaires
python-dateutil==2.8.2
pytz==2023.3

# Signature électronique (Yousign)
requests==2.31.0

# Automatisation
schedule==1.2.0

# Variables d'environnement
python-dotenv==1.0.0

# Traitement JSON
jsonschema==4.19.1
